wb_sys_dlg_font

+++ phpwb_sysdlg.c
@@ -227,7 +227,7 @@ ZEND_FUNCTION(wb_sys_dlg_font)
 
 	font = wbCreateFont((LPCTSTR)name, height, color, flags);
 	RETURN_LONG(
-		wbSysDlgFont(
+		(LONG)wbSysDlgFont(
 			(PWBOBJ)pwbparent,
 			 (LPTSTR)title,
 			  (PFONT)wbGetFont(font)
